+++ CHANGELOG.md
@@ -1,12 +1,17 @@
+## [0.2.2] - Date: 09-13-2021
+
+- Pack all the basic function to UniversalRouter class.
+- Added remove method to RouterInstance class.
+
 ## [0.2.1] - Date: 09-13-2021
 
-- Fixed problem of unchanging screen
+- Fixed problem of unchanging screen.
 
 
 ## [0.2.0] - Date: 09-12-2021
 
-- Optimized performance
-- Stop using the Provider package as a dependency
+- Optimized performance.
+- Stop using the Provider package as a dependency.
 
 
 ## [0.1.0] - Date: 08-24-2021